requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.8
django-allauth==0.33.0
django-crispy-forms==1.6.1
django-registration-redux==1.6
idna==2.6
oauthlib==2.0.2
psycopg2==2.7.3
python3-openid==3.1.0
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
